requirements.txt
antlr4-python3-runtime==4.9.3
anyio==4.2.0
astor==0.8.1
attrs==23.2.0
bandit==1.7.6
beautifulsoup4==4.12.2
black==24.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coverage==6.4.2
darglint==1.8.1
dependency-injector==4.41.0
docutils==0.20.1
dvc[ssh]==3.33.2
eradicate==2.3.0
fastapi==0.95.1
flake8==7.0.0
flake8-bandit==4.1.1
flake8-broken-line==1.0.0
flake8-bugbear==23.12.2
flake8-commas==2.1.0
flake8-comprehensions==3.14.0
flake8-debugger==4.1.2
flake8-docstrings==1.7.0
flake8-eradicate==1.5.0
flake8-isort==6.1.1
flake8-quotes==3.3.2
flake8-rst-docstrings==0.3.0
flake8-string-format==0.3.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
numpy==1.24.4
omegaconf==2.3.0
onnxruntime==1.16.3
opencv-python==4.6.0.66
packaging==22.0
pbr==6.0.0
pep8-naming==0.13.3
pillow==10.2.0
pluggy==1.3.0
py==1.11.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.11
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.17.2
pyparsing==3.1.1
pytest==7.1.2
pytest-cov==3.0.0
pygit2==1.13.3
python-multipart==0.0.5
PyYAML==6.0.1
requests==2.31.0
restructuredtext-lint==1.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
starlette==0.26.1
stevedore==5.1.0
testfixtures==6.18.5
tomli==2.0.1
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu
typing-extensions==4.9.0
urllib3==2.1.0
uvicorn==0.18.2
wemake-python-styleguide==0.18.0
